finish product

+++ ERPAnimalia/Controllers/VoucherController.cs
@@ -99,7 +99,7 @@ namespace ERPAnimalia.Controllers
         }
 
         [HttpGet]
-        public JsonResult GetProductDetail(int? page, int? limit, string term, int cantidad = 0, decimal descuento = 0)
+        public JsonResult GetProductDetail(int? page, int? limit, string term, int cantidad = 0, double descuento = 0)
         {      
             var detailGridTemp = TempData["DetailGrid"] as List<DetailGrid>;
             var detailGridList = new List<DetailGrid>();
@@ -182,7 +182,7 @@ namespace ERPAnimalia.Controllers
 
 
         [HttpGet]
-        public JsonResult GetSubtotal(Guid? idProduct, int cantidad = 0, decimal descuento = 0)
+        public JsonResult GetSubtotal(Guid? idProduct, int cantidad = 0, double descuento = 0)
         {
             var detailGridTemp = TempData["DetailGrid"] as List<DetailGrid>;
             var records = new List<DetailGrid>();
