+++ ERPAnimalia/Helper/CalCulos.cs
@@ -7,13 +7,13 @@ namespace ERPAnimalia.Helper
 {
     public class CalCulos
     {
-        public static string CalcularRentabilidad(decimal? precioCosto, decimal? PrecioVenta)
+        public static string CalcularRentabilidad(double? precioCosto, double? PrecioVenta)
         {
-            var pesos = (PrecioVenta.Value - precioCosto.Value).ToString("f2"); 
+            var pesos = (PrecioVenta.Value - precioCosto.Value).ToString("f2");            
             return pesos;
         }
 
-        public static string CalcularRentabilidadPorcentage(decimal precioCosto, decimal PrecioVenta)
+        public static string CalcularRentabilidadPorcentage(double precioCosto, double PrecioVenta)
         {
             var porcentage = (precioCosto / PrecioVenta).ToString("f2");
             return porcentage;
