+++ ERPAnimalia/Models/DetailGrid.cs
@@ -10,13 +10,13 @@ namespace ERPAnimalia.Models
      public Guid IdProduct { get; set; }
      public string Codigo { get; set; }
      public string Descripcion1 { get; set; }
-     public decimal PrecioVenta { get; set; }
+     public double PrecioVenta { get; set; }
      public decimal Cantidad { get; set; }
      public decimal kg { get; set; }
-     public decimal Descuento { get; set; }
+     public double Descuento { get; set; }
      public string Subtotal { get; set; }
      public decimal Total { get; set; }
-     public decimal Porcentage { get; set; }
+     public double Porcentage { get; set; }
      public int CategoryItem { get; set; }
      public int SubCategoryItem { get; set; }
     }
